Change structure of test files
